Add fetchCJS to webworker on desktop
