+++ packages/core/package.json
@@ -54,7 +54,7 @@
     "material-ui-popup-state": "^5.0.0",
     "rbush": "^3.0.1",
     "react-error-boundary": "^4.0.3",
-    "serialize-error": "^11.0.0",
+    "serialize-error": "^8.0.0",
     "shortid": "^2.2.13",
     "svg-path-generator": "^1.1.0"
   },
