+++ packages/product-core/package.json
@@ -48,7 +48,7 @@
     "copy-to-clipboard": "^3.3.1",
     "librpc-web-mod": "^1.0.0",
     "react-error-boundary": "^4.0.3",
-    "serialize-error": "^11.0.0",
+    "serialize-error": "^8.0.0",
     "shortid": "^2.2.15"
   },
   "peerDependencies": {
